Convert the remainder of character format icons to ansi
- Use common ColourBox control for Rip/Character colour selection
