+++ Source/Pablo/Formats/Character/Controls/ColourPad.cs
@@ -26,19 +26,6 @@ namespace Pablo.Formats.Character.Controls
 
 		#region ColourBox
 		
-		public abstract class ColourBox : Drawable
-		{
-			public abstract Color Color { get; set; }
-
-			protected override void OnPaint(PaintEventArgs pe)
-			{
-				base.OnPaint(pe);
-				var rect = new Rectangle(Size);
-				pe.Graphics.FillRectangle(Color, rect);
-				pe.Graphics.DrawButtonOutline(rect);
-			}
-		}
-		
 		class FBColourBox : ColourBox
 		{
 			public ColourPad Pad { get; set; }
