+++ Source/Pablo/Formats/Character/Tools/LineTool.cs
@@ -155,14 +155,12 @@ namespace Pablo.Formats.Character.Tools
 
 		Control ApplyColourButton()
 		{
-			var control = new ImageButton
+			var control = new AnsiButton
 			{
-				Image = ImageCache.BitmapFromResource("Pablo.Formats.Character.Icons.ApplyColour.png"),
+				Document = ImageCache.CharacterFromResource("Pablo.Formats.Character.Icons.ApplyColour.ans", false),
 				Toggle = true,
 				Pressed = ApplyColour,
-#if DESKTOP
 				ToolTip = "Draw with color"
-#endif
 			};
 
 			control.Click += delegate
